#290 Install sysbench from sources in the Docker image for ProxySQL
We need this because we want the .lua files for the test scenarios.
Also, I put it on the ProxySQL instance because the alternative would
have been to put it in the Vagrant machine that encapsulates the tests.
There were 2 disadvantages for that:
* you could not easily run the tests locally unless you were also using
Vagrant
* sometimes there are connectivity problems (due to the way in which
Docker forwards ports) which makes certain MySQL clients not work