T87 Avoid applying configs issued by self
Implemented mechanism that prevents a poxysql instance to apply the configs
that it just pushed to the cluster. This is needed because the local Consul
instance watches for all updates, including those comming from localhost.
A file is used to store a unique uuid per proxysql-consul instance and the uuid
is passed along with the config to identify the source.