Implementation of variable max_allowed_packet
This variable limits the amount of data that the client can send to the backend.
Note that ProxySQL won't really limit the amount of data that the client can send to the proxy, but the amount of data that the prox can send to the backend.

This commits also fixes issue #526

Closes #526