Upgrading MariaDB Connector C to 2.3.1
ProxySQL used a patched version of MariaDB Connector C 2.1.0 (aka libmariadbclient).
That version was patched to fix several bugs.
Although, version 2.1.0 seems to be quite broken in handling Prepared Statements, so now MariaDB Connector C was upgraded to version 2.3.1 .
Since this is a key compoenent in ProxySQL, several tests need to be performed because of this major change.