Renamed CPP to CXX in Makefiles
Renamed all variables in Makefiles that have 'CPP' in the name to 'CXX':

"CFLAGS enables the addition of switches for the C compiler, while
CXXFLAGS is meant to be used when invoking a C++ compiler. Similarly, a
variable CPPFLAGS exists with switches to be passed to the C or C++
preprocessor." (https://en.wikipedia.org/wiki/CFLAGS)

This change facilitates the building with customized CXXFLAGS (rpmbuild
exports a customized CXXFLAGS with hardening flags).