Remove hardcoded capabilities that the client uses
These hardcoded capabilities end up being set on all backend connections
ProxySQL makes. We don't want CLIENT_MULTI_RESULTS &
CLIENT_PS_MULTI_RESULTS to be enabled hardcoded, but they should be
dependent on the client connecting to ProxySQL.

Having this mismatch meant that any connection that was not setting
these flags would not be multiplexed with subsequent connections as the
client of client flags would not match.