use unit-of-work for consistent replica reads
This provides a way of ensuring that a series of database requests will
be forwarded to one server as long as a user-defined unit-of-work
identifier (or a series of such identifiers) remains the same across
the series of requests.  This is done by using the identifier(s),
received via query comment fields, to generate a numeric hash which is
ultimately used to index into the list of configured servers within a
hostgroup.

Also, the delimiters between comment fields and within key-value pairs
have been made configurable, with the standard ";" and "=" delimiters
established as defaults.