Change default value of mysql-session_idle_ms from 1000 to 1
Not only the default value is change from 1000 to 1 , but it also perform an upgrade from earlier release.
If during bootstrap a value of 1000 is found, it is automatically changed to 1 and a warning is generated.
This release also prevents the value 1000 to be a valid one during bootstrap.
This means that it is still possible to set the value to 1000 and to have a running proxysql instance with value 1000 , but during bootstrap it will be switched to 1.

If you really want to set mysql-session_idle_ms close to 1000 , it is recommended to set it to a closer value like 999 or 1001