Preserve epoch when synching with ProxySQL Cluster
On a large ProxySQL Cluster it was possible that an old configuration reaches
a node (nodeA) later than a new configuration is applied to another node (nodeB).
This has the side affect that nodeB will then believe that nodeA has a newer
configuration, while in reality is an old configuration applied later.

This commit preverves epoch, thus should avoid this problem