In Cluster pull_mysql_servers_from_peer() lock mysql_servers only if mysql servers tables are going to be written.
Before this commit we were calling GloAdmin->mysql_servers_wrlock() at the
beginning of the synchronization.
This was not ideal as we should call GloAdmin->mysql_servers_wrlock() only
if we are really going to write the tables. In other words, only if the
resultsets are fetched from the remote peer, and if their checksums didn't
change in the meantime.