fix NO_BACKSLASH_ESCAPE
Only `set_no_backslash_escapes(true)` when client actually wants to set the `NO_BACKSLASH_ESCAPE` flag in the sql-mode session variable - this probably was a typo initially. 

This erroneous behavior was detected when trying to run Rails with the latest proxysql 2.4.2 version - it returned this slightly misleading `ArgumentError: negative string size (or size too big)` error message during startup, that seems to have to do with libmysql's `escape` functionality. With proxysql 2.3.2 it works just fine.

BTW: Not sure why it is `NO_BACKSLASH_ESCAPE` instead of `NO_BACKSLASH_ESCAPES`, but I guess this doesn't really matter in this comparison.