Fix sequence id in EOF to OK packet conversion #3698
Fix invalid resultset when using query cache and mixing clients using
and not using CLIENT_DEPRECATE_EOF.

If a client connects to ProxySQL not using CLIENT_DEPRECATE_EOF and it
caches a resultset, when a client using CLIENT_DEPRECATE_EOF executes
the same query it will get an invalid resultset and the client will
disconnect. The data in the resultset is correct, but proxysql skips a
sequence id thus the client assumes it is corrupted.