Add test checking COM_REGISTER_SLAVE enables fast_forward
This test runs a custom binlog client, that uses libslave, to test that
ProxySQL activates fast forward automatically after receiving the
command, and to test that it can read the binlog events. The custom
client reads the binlog after sending a query that do not disables
multiplexing and after sending a query that disables multiplexing.