Add test checking binlog reader uses previous hostgroup
When a COM_REGISTER_SLAVE command is received, test that ProxySQL will
automatically switch from not fast_forward mode to fast_forward mode.
It also test that the destination hostgroup assigned from previous
COM_QUERY commands is the one used to establish the fast_forward
connection. To test this we look at how many connections are closed in
the hostgroup that should have been used for the fast_forward
connections.