Fix preservation of SHUNNED servers for GR actions
This commit adds support for preservation of SHUNNED servers during
hostgroup movements for GR actions. This is not going to be added to
current implementation due to the current mistmatch between SHUNNED
state for 'myghm.mysql_servers' table and the runtime memory status of
the server. This work is left as documentation for a potential future
implementation that doesn't use 'myghm.mysql_servers' as the source of
truth for performing the converge of configuration on GR monitoring
actions.