Lock on hostgroup only when session tracking fails
When unable to parse a set statement, and if variables
SQL_SESSION_TRACK_STATE_CHANGE and SQL_SESSION_TRACK_SYSTEM_VARIABLES
are correctly set, ProxySQL will send the whole SET statement to the
backend. If the backend will return a list of variables with only
variables that ProxySQL parses, lock on hostgroup will not be enabled.
Otherwise lock on hostgroup will be enabled