Add variable try_to_unlock_with_session_tracking
When a SET statement parse failure occurs, lock on hostgroup is enabled
and try_to_unlock_with_session_tracking variable is set to true. Due to
this, during processing of the backend response, if the session tracking
data only contains changes in variables that ProxySQL tracks, we remove
the lock on the hostgroup.

Also add some minor improvements and fixes to session tracking.