POC for hardcoded frontend/backend user mapping
This commit is just a POC!
It hardcodes the mapping of frontend to backend usernames.
Client can connect to ProxySQL with frontend credentials (username and password)
and in case of successful authentication it will replace username with the mapped
username. It will later retrieve the password using the backend username.

The mapping (and also the client authentication) will fail if the mapped username
doesn't exist, or if the password of the mapped username is not in clear text.

Because the mapped password must be in clear text, variable `admin-hash_passwords`
must be set to `false`.

Attributes like `default_hostgroup`, `default_schema`, and `transaction_persistent`
are all retrieved from the frontend user entry.

Mapping is not supported for ``COM_CHANGE_USER()` executed by client.