Fix servers resultsets not updated for empty config
If these resultsets are not updated, checksums will always mismatch
between the peer cluster nodes in case servers are deleted. The primary
will hold the correct checksum ('0x0'), but will report the old resultset
to the peer node, the second will compute a checksum different than the
expected '0x0'. Cycle won't stop until config is updated in the target
primary.