Clear SSL error queue on shutdown
SSL_shutdown can return an error and if it is not cleared it will be
left on the thread's error queue since they are thread local causing
unrelated connections running on the same thread to be incorrectly terminated.
Clear the error queue now if SSL_shutdown returns an error to keep us from
impacting other connections.