Fix 'test_backend_conn_ping-t' conn-pool cleanup
Since the test stress test the connection pool, verifying the limits of
backend connections that ProxySQL can maintain, it's possible to
propagate a connection pool filled with idle connections to the next
tests. This could lead to connection exhaustion on those tests. The
scenario could be prevented in a generic way, as part of the
reconfiguration between tests. It's left as a potential 'TODO' in the
test itself.