Adding some timers for debugging purposes
Adding Timers in Base_Thread to count the number of time and
elapsed time of when Session::handler() and Connection::handler()
are called.
This feature is disabled by default.

Timers are enabled only if in proxysql_macros.h the following
is defined: #define ENABLE_TIMER true

The default: #define ENABLE_TIMER false