Fix SSL error queue cleanup for backend conns
The SSL error queue wasn't cleanup after an SSL related error took place
in a backend connection. This would result in the propagation of the
error to other conns handled by the thread, which could result in:

- Incorrect destruction of connections in conn-pool.
- Invalid error propagation to clients.

This is a consequence of 'libmariadbclient' not performing a cleanup of
this queue by itself. The situation got mitigated since the library
**does** perform a cleanup of such queue during connect phase
('auth_caching_sha2_client|auth_sha256_client'), and ProxySQL does a
cleanup of this queue during frontend SSL traffic.