Fix value on first fetch for connpool prometheus metrics
The initial value for metrics using 'p_update_connection_pool_update_*'
helpers were missing. This only affects to the first metrics fetch,
during the metric creation on the internal maps, subsequent fetches will
find the already created metric and updated the value properly. As
real counters are internal, this doesn't affect the correctness of the
value in subsequent fetches.